Fixes for Chebyshev and uniform API for Bessel and Butterworth

+++ ATK/EQ/BesselFilter.cpp
@@ -9,7 +9,7 @@
 namespace
 {
   template<typename DataType>
-  void create_bessel_analog_coefficients(size_t order, std::vector<std::complex<DataType> >& z, std::vector<std::complex<DataType> >& p, DataType& k)
+  void create_bessel_analog_coefficients(int order, std::vector<std::complex<DataType> >& z, std::vector<std::complex<DataType> >& p, DataType& k)
   {
     k = 1;
     z.clear(); // no zeros for this filter type
@@ -148,7 +148,7 @@ namespace
     DataType k;
     
     int fs = 2;
-    create_bessel_analog_coefficients(order, z, p, k);
+    create_bessel_analog_coefficients(static_cast<int>(order), z, p, k);
     DataType warped = 2 * fs * std::tan(boost::math::constants::pi<DataType>() *  Wn / fs);
     zpk_lp2lp(warped, z, p, k);
     zpk_bilinear(fs, z, p, k);
@@ -178,7 +178,7 @@ namespace
     DataType k;
     
     int fs = 2;
-    create_bessel_analog_coefficients(order/2, z, p, k);
+    create_bessel_analog_coefficients(static_cast<int>(order/2), z, p, k);
     wc1 = 2 * fs * std::tan(boost::math::constants::pi<DataType>() * wc1 / fs);
     wc2 = 2 * fs * std::tan(boost::math::constants::pi<DataType>() * wc2 / fs);
     
@@ -210,7 +210,7 @@ namespace
     DataType k;
     
     int fs = 2;
-    create_bessel_analog_coefficients(order/2, z, p, k);
+    create_bessel_analog_coefficients(static_cast<int>(order/2), z, p, k);
     wc1 = 2 * fs * std::tan(boost::math::constants::pi<DataType>() * wc1 / fs);
     wc2 = 2 * fs * std::tan(boost::math::constants::pi<DataType>() * wc2 / fs);
